# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/media": {
      "get": {
        "operationId": "MediaController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Media"
        ]
      },
      "post": {
        "operationId": "MediaController_upload",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Media"
        ]
      }
    },
    "/admin/media/{id}": {
      "patch": {
        "operationId": "MediaController_setAltText",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAltTextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Media"
        ]
      },
      "delete": {
        "operationId": "MediaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Media"
        ]
      }
    },
...
    "/admin/testimonials/{id}/status": {
      "patch": {
        "operationId": "AdminTestimonialsController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTestimonialStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminTestimonials"
        ]
      }
    },
    "/admin/testimonials/{id}/feature": {
      "patch": {
        "operationId": "AdminTestimonialsController_setFeature",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTestimonialFeatureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminTestimonials"
        ]
      }
    },
    "/admin/testimonials": {
      "get": {
        "operationId": "AdminTestimonialsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "published",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminTestimonials"
        ]
      },
      "post": {
        "operationId": "AdminTestimonialsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/admin_testimonialsCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminTestimonials"
        ]
      }
    },
    "/admin/testimonials/{id}": {
      "get": {
        "operationId": "AdminTestimonialsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminTestimonials"
        ]
      },
      "patch": {
        "operationId": "AdminTestimonialsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/admin_testimonialsUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminTestimonials"
        ]
      },
      "delete": {
        "operationId": "AdminTestimonialsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "AdminTestimonials"
        ]
      }
    },
...
  },
  "info": {
    "title": "Safeer API",
    "description": "جمعية سفير الدعوية / Safeer Association — content and admin API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sf_sid"
      }
    },
    "schemas": {
...
      "SetAltTextDto": {
        "type": "object",
        "properties": {
          "altAr": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "altEn": {
            "type": "string",
            "maxLength": 255,
            "nullable": true
          }
        },
        "required": [
          "altAr"
        ],
        "additionalProperties": false
      },
...
      "SetTestimonialStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "published",
              "hidden"
            ]
          }
        },
        "required": [
          "status"
        ],
        "additionalProperties": false
      },
      "SetTestimonialFeatureDto": {
        "type": "object",
        "properties": {
          "isFeatured": {
            "type": "boolean"
          }
        },
        "required": [
          "isFeatured"
        ],
        "additionalProperties": false
      },
      "admin_testimonialsCreateDto": {
        "type": "object",
        "properties": {
          "quoteAr": {
            "type": "string",
            "minLength": 1
          },
          "quoteEn": {
            "type": "string",
            "nullable": true
          },
          "authorName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 191
          },
          "authorDescAr": {
            "type": "string",
            "maxLength": 255,
            "nullable": true
          },
          "authorDescEn": {
            "type": "string",
            "maxLength": 255,
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "published",
              "hidden"
            ]
          },
          "isFeatured": {
            "type": "boolean"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "contact_form"
            ]
          },
          "sourceMessageId": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "quoteAr",
          "authorName"
        ],
        "additionalProperties": false
      },
...
      "admin_testimonialsUpdateDto": {
        "type": "object",
        "properties": {
          "quoteAr": {
            "type": "string",
            "minLength": 1
          },
          "quoteEn": {
            "type": "string",
            "nullable": true
          },
          "authorName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 191
          },
          "authorDescAr": {
            "type": "string",
            "maxLength": 255,
            "nullable": true
          },
          "authorDescEn": {
            "type": "string",
            "maxLength": 255,
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "published",
              "hidden"
            ]
          },
          "isFeatured": {
            "type": "boolean"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "contact_form"
            ]
          },
          "sourceMessageId": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}